created tool package + splitted datasetCreator to NN package
